refactor border for title, textStyle and label close #67

+++ ggdsl-echarts/src/main/kotlin/org/jetbrains/kotlinx/ggdsl/echarts/translator/option/EchartsTitle.kt
@@ -33,7 +33,7 @@ internal data class EchartsTitle(
     val bottom: @Contextual Any? = null,
     val backgroundColor: EchartsColor? = null,
     val borderColor: EchartsColor? = null,
-    val borderWidth: Int? = null,
+    val borderWidth: Double? = null,
     val borderRadius: @Contextual Any? = null, // number, array
     val shadowBlur: Int? = null,
     val shadowColor: EchartsColor? = null,
